server/pyproject.toml
[tool.poetry]
name = "plummycodechat"
version = "0.1.0"
description = "plummycode's chat"
authors = ["Andrew Glushnev <glushnev22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.4"
asgiref = "^3.7.2"
Django = "^4.2.4"
django-jazzmin = "^2.6.0"
djangorestframework = "^3.14.0"
mysqlclient = "2.2.0"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
simplejwt = "^2.0.1"
sqlparse = "^0.4.4"
typing = "^3.7.4.3"
typing_extensions = "^4.7.1"
tzdata = "^2023.3"
celery = {extras = ["redis"], version = "^5.3.1"}
gunicorn = "^21.2.0"
channels = {extras = ["daphne"], version = "^4.0.0"}
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^4.2.0"
drf-spectacular = "^0.26.4"
sorl-thumbnail = "^12.9.0"
pillow = "^10.0.0"
django-filter = "^23.2"




[tool.poetry.group.dev.dependencies]
pylint-django = "^2.5.3"
isort = "^5.12.0"
autopep8 = "^2.0.2"
pytest-django = "^4.5.2"
pylint-celery = "^0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
